requirements.txt
discord.py
python-dotenv
yt-dlp
pynacl
ffmpeg
flask
waitress
aiohttp
re